# lib/resupply-api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title — orval uses it to name generated output files.
  title: Api
  version: 0.1.0
  description: |
    CPAP Resupply Automation admin API. Powers the admin dashboard
    (queue, episode detail, conversation viewer, fulfillment). Patient
    PHI flows through this API and lives in the resupply.* schema.
servers:
  - url: /resupply-api
    description: Base API path
...
paths:
...
  /patients:
    get:
      operationId: listPatients
      tags: [patients]
      summary: Paginated patient list
      description: |
        Paginated list of patients. Decrypts firstName + lastName for
        display. Phone and email are surfaced as booleans only —
        `hasPhone` / `hasEmail` — so admins can tell which channels
        are reachable without leaking the values themselves.

        `search` matches against pacwareId (plaintext, indexed) AND
        against the decrypted first/last name (server-side `pgp_sym_decrypt`).
        Decrypted-name search is a Postgres-side full scan; the limit
        cap and the admin-only gate make this acceptable for the
        small steady-state admin workload.
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [active, paused, closed]
          description: Filter by patient status.
        - in: query
          name: search
          schema:
            type: string
            maxLength: 64
          description: |
            Case-insensitive substring match against pacwareId or the
            decrypted first/last name.
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: Paginated page of patients.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientListPage"
        "400":
          description: Query string failed validation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsoleValidationError"
        "401":
          description: No session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthError"
        "403":
          description: Caller is not on the admin allowlist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthError"
    post:
      operationId: createPatient
      tags: [patients]
      summary: Create a new patient
      description: |
        Admin-initiated patient creation from the dashboard. Encrypts
        first/last name, DOB, phone, email, and address at the SQL
        site via pgcrypto. Returns the new patient's id; the
        dashboard navigates to /patients/{id} on success.

        409 with `error: "duplicate_pacware_id"` when the supplied
        Pacware id is already in use — pacware_id is the unique join
        key for CSV imports and may not be reused.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePatientRequest"
      responses:
        "201":
          description: Patient created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatePatientResponse"
        "400":
          description: Body failed validation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsoleValidationError"
        "401":
          description: No session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthError"
        "403":
          description: Caller is not on the admin allowlist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthError"
        "409":
          description: Pacware id already in use.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DuplicatePacwareIdError"
...
  /patients/import-csv:
    post:
      operationId: importPatientsCsv
      tags: [patients]
      summary: Bulk-create patients from a parsed Pacware-style export
      description: |
        Accepts a JSON array of up to 500 already-parsed CSV rows.
        The dashboard parses the file client-side (papaparse),
        validates per-row, and POSTs JSON. Each row is encrypted
        and inserted in its own try/catch so a single bad row
        cannot tank the batch. Duplicates on `pacware_id` are
        counted as `skippedDuplicates`, not errors. ONE audit row
        per call records structural counts only — no PHI.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportPatientsCsvRequest"
      responses:
        "200":
          description: Batch processed; per-row outcomes returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportPatientsCsvResponse"
        "400":
          description: Body failed validation (e.g. > 500 rows).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsoleValidationError"
        "401":
          description: No session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthError"
        "403":
          description: Caller is not on the admin allowlist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthError"
components:
  schemas:
...
    AuthError:
      type: object
      description: |
        Auth failure body. Carries a single human-readable `error`
        string safe for the dashboard to render verbatim. Never
        contains the underlying error message, the user id, or any
        connection details.
      properties:
        error:
          type: string
      required:
        - error
...
    ConsoleValidationError:
      type: object
      description: Returned when a list endpoint's query string fails zod validation.
      properties:
        error:
          type: string
          enum: [invalid_query]
        issues:
          type: array
          items:
            type: object
            properties:
              path:
                type: string
              message:
                type: string
            required: [path, message]
      required:
        - error
        - issues
...
    DuplicatePacwareIdError:
      type: object
      description: |
        Returned by POST /patients when the supplied pacwareId is
        already in use. The dashboard renders the message inline on
        the create-patient form so the admin can correct or jump to
        the existing record.
      properties:
        error:
          type: string
          enum: [duplicate_pacware_id]
        message:
          type: string
      required:
        - error
        - message
    CreatePatientRequest:
      type: object
      description: |
        Body for POST /patients. PHI fields (name, DOB, phone, email,
        address) are pgcrypto-encrypted server-side at the SQL site —
        the API never logs them and the request body never echoes back
        in any response.
      properties:
        pacwareId:
          type: string
          minLength: 1
          maxLength: 64
          description: |
            Pacware patient id. Plaintext, indexed, unique. The same
            id used by the legacy DME system; required so future CSV
            imports correlate cleanly.
        legalFirstName:
          type: string
          minLength: 1
          maxLength: 80
        legalLastName:
          type: string
          minLength: 1
          maxLength: 80
        dateOfBirth:
          type: string
          pattern: "^\\d{4}-\\d{2}-\\d{2}$"
          description: ISO date in YYYY-MM-DD form (no time / timezone).
        phoneE164:
          type: string
          nullable: true
          pattern: "^\\+[1-9]\\d{7,14}$"
          description: |
            E.164 phone (e.g. `+14155551212`). Null / omitted means
            the patient has no phone on file; SMS / voice reminders
            will not be attempted.
        email:
          type: string
          format: email
          nullable: true
          maxLength: 254
        address:
          nullable: true
          type: object
          properties:
            line1: { type: string, minLength: 1, maxLength: 120 }
            line2: { type: string, maxLength: 120 }
            city: { type: string, minLength: 1, maxLength: 80 }
            state: { type: string, minLength: 1, maxLength: 40 }
            postalCode: { type: string, minLength: 1, maxLength: 20 }
            country: { type: string, minLength: 1, maxLength: 40 }
          required: [line1, city, state, postalCode, country]
        status:
          type: string
          enum: [active, paused, closed]
          description: Defaults to `active` when omitted.
        insurancePayer:
          type: string
          nullable: true
          maxLength: 120
          description: |
            Free-text payer name (e.g. "Aetna"). Used as a match
            target by frequency_rules.
        cadenceOverrideDays:
          type: integer
          nullable: true
          minimum: 1
          maximum: 365
          description: |
            Per-patient cadence override (days). Wins over rules and
            the prescription cadence when set at create time.
        channelPreference:
          type: string
          enum: [sms, email, voice]
          nullable: true
          description: Per-patient outbound channel override.
      required:
        - pacwareId
        - legalFirstName
        - legalLastName
        - dateOfBirth
    CreatePatientResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
      required:
        - id
...
    PatientListItem:
      type: object
      description: |
        One row in the patient list. Decrypts firstName + lastName for
        display. NEVER includes the phone number or email — only booleans
        showing whether the patient has each on file. The lastMessage*
        triplet is sourced from the patient_latest_message projection
        (refreshed in-line on every inbound/outbound message write) so
        the list can show "last contacted" without a per-row scan of
        the messages table.
      properties:
        id:
          type: string
          format: uuid
        pacwareId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        status:
          type: string
          enum: [active, paused, closed]
        hasPhone:
          type: boolean
        hasEmail:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        lastMessageAt:
          type: string
          format: date-time
          nullable: true
          description: |
            Timestamp of the most recent inbound or outbound message
            for this patient across any conversation. NULL when the
            patient has never exchanged a message.
        lastMessageDirection:
          type: string
          enum: [inbound, outbound]
          nullable: true
          description: |
            Direction of that most recent message. NULL when there is
            no message yet.
        lastMessagePreview:
          type: string
          nullable: true
          description: |
            Decrypted, single-line preview of the most recent message
            body, truncated to 80 characters by the projection helper.
            NULL when there is no message yet. PHI: only render in
            authenticated admin contexts.
      required:
        - id
        - pacwareId
        - firstName
        - lastName
        - status
        - hasPhone
        - hasEmail
        - createdAt
        - updatedAt
    PatientListPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/PatientListItem"
        total:
          type: integer
          minimum: 0
        limit:
          type: integer
          minimum: 1
          maximum: 100
        offset:
          type: integer
          minimum: 0
      required:
        - items
        - total
        - limit
        - offset
...
    ImportPatientRow:
      type: object
      description: |
        One CSV row, post-parse. Must satisfy the same per-field
        rules as POST /patients (E.164 phone, ISO date of birth,
        all-or-none address). Address line2 / country are optional
        even when the rest of the address is provided.
      properties:
        pacwareId:
          type: string
          minLength: 1
          maxLength: 64
        legalFirstName:
          type: string
          minLength: 1
          maxLength: 80
        legalLastName:
          type: string
          minLength: 1
          maxLength: 80
        dateOfBirth:
          type: string
          pattern: "^\\d{4}-\\d{2}-\\d{2}$"
        phoneE164:
          type: string
          pattern: "^\\+[1-9]\\d{7,14}$"
        email:
          type: string
          format: email
          maxLength: 254
        addressLine1:
          type: string
          maxLength: 160
        addressLine2:
          type: string
          maxLength: 160
        city:
          type: string
          maxLength: 80
        state:
          type: string
          maxLength: 40
        postalCode:
          type: string
          maxLength: 20
        country:
          type: string
          maxLength: 40
        insurancePayer:
          type: string
          maxLength: 120
      required:
        - pacwareId
        - legalFirstName
        - legalLastName
        - dateOfBirth
      additionalProperties: false
    ImportPatientsCsvRequest:
      type: object
      description: |
        Wrapped row array. Server caps batch size at 500 to keep
        the request well under the body-parser timeout; the
        dashboard chunks larger imports into multiple requests.
      properties:
        rows:
          type: array
          minItems: 1
          maxItems: 500
          items:
            $ref: "#/components/schemas/ImportPatientRow"
      required:
        - rows
      additionalProperties: false
    ImportPatientRowError:
      type: object
      properties:
        rowIndex:
          type: integer
          minimum: 0
        field:
          type: string
        message:
          type: string
      required:
        - rowIndex
        - message
    ImportPatientsCsvResponse:
      type: object
      description: |
        Per-call import result. `errors[]` lets the dashboard render
        a downloadable error CSV; `createdIds[]` is empty for
        skipped/error rows so the dashboard can compute "I
        successfully created N patients in this batch".
      properties:
        created:
          type: integer
          minimum: 0
        skippedDuplicates:
          type: integer
          minimum: 0
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ImportPatientRowError"
        createdIds:
          type: array
          items:
            type: string
            format: uuid
      required:
        - created
        - skippedDuplicates
        - errors
        - createdIds